% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci.R
\name{add_ci}
\alias{add_ci}
\title{Add bootstrapped confidence intervals}
\usage{
add_ci(x, ...)
}
\arguments{
\item{x}{`sm_predict` or `sm_regression` object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function is meant to be run after `sm_predict` or `sm_regression` and
will calculate bootstrapped confidence intervals for `sm_predict` objects
and add the bootstraped model estimates for `sm_regression` objects.
}
\seealso{
\code{\link{add_ci.sm_regression}}
}
