% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coef.R
\name{add_coef}
\alias{add_coef}
\title{Grab coefficient from models}
\usage{
add_coef(x)
}
\arguments{
\item{x}{`sm_regression` object}
}
\description{
This function is meant to be run after `sm_regression` when the regression model
is univariate.  From each of the weighted models, the coefficient will be
saved as a column in the output data frame.
}
\examples{
sm_regression(
  data = mtcars,
  method = "lm",
  formula = mpg ~ am ,
  weighting_var = "hp",
  newdata = data.frame(hp = c(150, 200))
) \%>\%
add_coef()
}
\seealso{
\code{\link{sm_regression}}
}
