% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coef_sd.R
\name{add_coef_sd}
\alias{add_coef_sd}
\title{Estimate the SD for model coefficients}
\usage{
add_coef_sd(x)
}
\arguments{
\item{x}{`sm_regression` object
is `FALSE`}
}
\description{
This function is meant to be run after `sm_regression() %>% add_ci()`
when the regression model is univariate.  From each of the weighted models,
the coefficient standard deviation will be estimated and
saved as a column in the output data frame.
}
\seealso{
\code{\link{sm_regression}}
}
