% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculate_weights.R
\name{caclulate_weights}
\alias{caclulate_weights}
\title{Provided a point, a data frame, and lambda, this function calculates the
weights each observation in the data frame provides the calculations}
\usage{
caclulate_weights(data, point, kernel, dist_method, lambda, covars)
}
\arguments{
\item{data}{data frame}

\item{point}{center point around which weights are calculated in data frame}

\item{kernel}{type of kernel to use for weight calculation}

\item{dist_method}{the distance measure to be used.
This must be one of "euclidean", "maximum", "manhattan", "canberra",
"binary" or "minkowski". Any unambiguous substring can be given.}

\item{lambda}{lambda parameter in kernel.
Often the circumference of points with positive weights}

\item{covars}{character vector of column names of the covariates}
}
\description{
Provided a point, a data frame, and lambda, this function calculates the
weights each observation in the data frame provides the calculations
}
\keyword{internal}
