% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculate_dist.R
\name{calculate_dist}
\alias{calculate_dist}
\title{Calculate the distance vector between a single point and each observation
in the data frame}
\usage{
calculate_dist(data, point, dist_method)
}
\arguments{
\item{data}{data frame}

\item{point}{center point around which weights are calculated in data frame}

\item{dist_method}{the distance measure to be used.
This must be one of "euclidean", "maximum", "manhattan", "canberra",
"binary" or "minkowski". Any unambiguous substring can be given.}
}
\description{
Calculate the distance vector between a single point and each observation
in the data frame
}
\keyword{internal}
