% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculate_weights.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Provided a point, a data frame, and lambda, this function calculates the
weights each observation in the data frame provides the calculations}
\usage{
calculate_weights(dist, lambda, kernel, weighting_var)
}
\arguments{
\item{dist}{distance vector}

\item{lambda}{The radius of the kernel for tri-cubic, Epanechnikov, and flat kernels.
The standard deviation for the Gaussian kernel.}

\item{kernel}{type of kernel to use for weight calculation}
}
\description{
Provided a point, a data frame, and lambda, this function calculates the
weights each observation in the data frame provides the calculations
}
\keyword{internal}
