% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-message_print.R
\name{message_print}
\alias{message_print}
\title{Prints errors/warnings/messages}
\usage{
message_print(data, message_col, preamble)
}
\arguments{
\item{data}{data frame}

\item{message_col}{string with column name of message to print}

\item{preamble}{character string to print before the message
(e.g. 'Error from coxph:')}
}
\description{
Prints errors/warnings/messages from the regression model function and from the
predict function.  The function will print each unique message along with the
data associated with the message.
}
