% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sjosmooth.df}
\alias{sjosmooth.df}
\title{Same as sjosmooth.tbl, but as a data frame.  Primarily used for testing.
A simulated data frame used to motivate examples in the \code{sjosmooth} package}
\format{A data frame with 7 variables:
\describe{
  \item{x}{Covariate x}
  \item{y}{Covariate y}
  \item{xb}{Linear portion of Cox regression model (i.e. x + y)}
  \item{time.true}{The true time to death}
  \item{time.cens}{Censoring time}
  \item{time.obs}{Observed time to death/censoring time}
  \item{event}{Death observed: yes/no (1/0)}
}}
\usage{
sjosmooth.df
}
\description{
Same as sjosmooth.tbl, but as a data frame.  Primarily used for testing.
A simulated data frame used to motivate examples in the \code{sjosmooth} package
}
\keyword{datasets}
