% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjosmooth.R
\docType{package}
\name{sjosmooth}
\alias{sjosmooth}
\alias{sjosmooth-package}
\title{sjosmooth: A package for perform n-dimensional kernel smoothing for Cox proportional hazards regression models.}
\description{
sjosmooth: A package for perform n-dimensional kernel smoothing for Cox proportional hazards regression models.
}
\section{sjosmooth functions}{

sm.timeto
}

