% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.coxph.R
\name{sm.coxph}
\alias{sm.coxph}
\title{Univariate and multivariable kernel smoothing.}
\usage{
sm.coxph(formula, data, newdata = NULL, time = NULL, bandwidth = 0.8,
  lambda = NULL, type = c("survival", "failure", "expected", "median",
  "quantile", "risk", "lp"), kernel = c("epanechnikov", "tricube", "gaussian",
  "flat"), dist.method = "euclidean", quantile = NULL, ...)
}
\arguments{
\item{formula}{The formula call for the model. For exampe, \code{survival::coxph(formula = Surv(time, dead) ~ x + y)}.
Do not include transformed variables in the formula (e.g. \code{Surv(time, dead) ~ log(x) + y})--the specified
\code{lambda} will no longer be interpretable.}

\item{data}{A data.frame or tibble with named columns.}

\item{newdata}{A data.frame or tibble with named columns.
Predictions will only be estimated at these points, if specified.}

\item{time}{Specifies the timepoint where kernel smoothing will be performed.  Applies to \code{type = c("survival", "failure", "expected")}
If time is not specified, then the time from data (or newdata) will be used.}

\item{bandwidth}{The proportion of data to be included in each kernel-smoothed estimate.  Univariate models only.}

\item{lambda}{The radius of the kernel for tri-cubic, Epanechnikov, and flat kernels.
The standard deviation for the Gaussian kernel.}

\item{type}{Specifies the type of statistic that will be calculated.  Default is \code{survival}.}

\item{kernel}{Specifies the kernel to be used: \code{epanechnikov}, \code{tricube}, \code{gaussian}, and \code{flat} are accepted.
Default is \code{epanechnikov}.  See vignettes for differences in methods.}

\item{dist.method}{Specifies the distance measure to be used in the kernel.  Default is \code{euclidean}.
Any distance measure accepted by \code{stats::dist} is acceptable.}

\item{quantile}{If \code{type = "quantile"}, specify the quantile to be estimates with a number between 0 and 1.}

\item{...}{Additional arguments to be passed to survival::coxph function call.}
}
\value{
A vector with the estimated (i.e. kernel-smoothed) outcomes.
}
\description{
The purpose of this function is to provide kernel-smoothed
estimates for n-dimensional data. For example, using two predictors, x and y,
to predict an outcome. The program provides an estimated hyperplane of the outcome by x and y.
}
