% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_cont_interaction.R
\name{sm_beta_interaction}
\alias{sm_beta_interaction}
\title{Illustrate interactions by estimating changes in beta coefficient(s) over
a continuous variable.}
\usage{
sm_beta_interaction(data, method, formula, interaction, newdata = data,
  method.args = NULL, lambda = 1, exponentiate = FALSE)
}
\arguments{
\item{data}{data frame}

\item{method}{(function) to use}

\item{formula}{formula}

\item{interaction}{third variable interaction term}

\item{newdata}{new data frame.  Default is `data`.  Only requires third
term from `interaction`}

\item{method.args}{List of additional arguments passed on to the
modelling function defined by `method`}

\item{lambda}{The radius of the kernel for tri-cubic, Epanechnikov, and flat kernels.
The standard deviation for the Gaussian kernel}

\item{exponentiate}{Default is `FALSE`.  Exponentiate beta coefficients}
}
\description{
This function illustrates how the relationship between an independent
variable and a dependent variable is modified over the values of a third
variable.  Using `data`, `method`, and `formula` the regression model is
estimated; the estimates are weighted by the third `interaction`
variable.
}
\examples{
sm_beta_interaction(
  data = mtcars,
  method = "lm",
  formula = mpg ~ am ,
  interaction = "hp"
)
}
