% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sm_predict.R
\name{sm_predict}
\alias{sm_predict}
\alias{sm_predict_safely}
\title{Wrapper for `predict()` with custom coding for some types of predictions.}
\usage{
sm_predict(method, object, newdata, type)

sm_predict_safely(...)
}
\arguments{
\item{method}{type of regression model}

\item{object}{the results of a regression fit}

\item{newdata}{New data at which to do predictions}

\item{type}{the type of predicted value}
}
\description{
Wrapper for `predict()` with custom coding for some types of predictions.
}
\keyword{internal}
