% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sm_predict_raw.R
\name{sm_predict_raw}
\alias{sm_predict_raw}
\title{Wrapper for `predict()` with custom coding for some types of predictions.}
\usage{
sm_predict_raw(method, object, newdata, type, conf.level = 0.95)
}
\arguments{
\item{method}{type of regression model}

\item{object}{the results of a regression fit}

\item{newdata}{New data at which to do predictions}

\item{type}{the type of predicted value}

\item{conf.level}{The confidence level to use for the confidence interval.
Must be strictly greater than 0 and less than 1. Default is 0.95.}
}
\description{
Wrapper for `predict()` with custom coding for some types of predictions.
}
\keyword{internal}
