% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_cont_interaction.R
\name{sm_wt_regression}
\alias{sm_wt_regression}
\title{Kernel-weighted regression modelling}
\usage{
sm_wt_regression(data, method, formula, weighting_var, newdata = data,
  method.args = NULL, lambda = 1)
}
\arguments{
\item{data}{data frame}

\item{method}{(function) to use}

\item{formula}{formula}

\item{weighting_var}{columns name(s) of variables used to calculate weights}

\item{newdata}{new data frame.  Default is `data`.  Only requires third
term from `interaction`}

\item{method.args}{List of additional arguments passed on to the
modelling function defined by `method`}

\item{lambda}{The radius of the kernel for tri-cubic, Epanechnikov, and flat kernels.
The standard deviation for the Gaussian kernel}
}
\description{
General function that fits weighted regerssion models, where the weights
are calculated from a ancillary variable(s) or from variable(s) found in the
regression model. Using `data`, `method`, and `formula` the regression model is
estimated; the estimates are weighted by the variable(s) listed in `weighting_var`.
}
\examples{
sm_wt_regression(
  data = mtcars,
  method = "lm",
  formula = mpg ~ am ,
  weighting_var = "hp",
  lambda = 2
)
}
