% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci.R
\name{add_ci.sm_regression}
\alias{add_ci.sm_regression}
\title{Add bootstrapped model estimates}
\usage{
\method{add_ci}{sm_regression}(x, bootn = 200, ...)
}
\arguments{
\item{x}{`sm_regression` object}

\item{bootn}{number of bootstrap model to run. Default is 200.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function is meant to be run after `sm_regression` and
will calculate bootstraped model estimates.
}
\examples{
sm_regression(
  data = mtcars,
  method = "lm",
  formula = mpg ~ am ,
  weighting_var = "hp",
  lambda = 1,
  newdata = data.frame(hp = c(150, 200))
) \%>\%
add_ci(bootn = 50)
}
