% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coef.R
\name{add_coef}
\alias{add_coef}
\title{Grab coefficient from models}
\usage{
add_coef(x)
}
\arguments{
\item{x}{`sm_regression` object}
}
\description{
This function is meant to be run after `sm_regression` when the regression model
is univariate.  From each of the weighted models, the coefficient will be
saved as a column in the output data frame.
}
\seealso{
\code{\link{sm_regression}}
}
