% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjosmooth-package.R
\docType{package}
\name{sjosmooth-package}
\alias{sjosmooth}
\alias{sjosmooth-package}
\title{sjosmooth: Kernel smoothing for time to event data}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

The sjosmooth package was built to calculate
    kernel-weighted estimates from time-to-event endpoints.  Importantly,
    the package allows for 1 or more weighting variables, allowing for the
    estimation of an n-dimensional hyperplane.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.danieldsjoberg.com/sjosmooth/}
  \item Report bugs at \url{https://github.com/ddsjoberg/sjosmooth/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danieldsjoberg@gmail.com}

}
\keyword{internal}
