% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjosmooth-package.R
\docType{package}
\name{sjosmooth-package}
\alias{sjosmooth}
\alias{sjosmooth-package}
\title{sjosmooth: Kernel smoothing for time to event data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danieldsjoberg@gmail.com}

}
\keyword{internal}
