% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.coxph.R
\name{sm.coxph}
\alias{sm.coxph}
\title{n-dimensional kernel-smoothed survival probabilites.}
\usage{
sm.coxph(formula, data, pred.time = 1, lambda = 1, grid = NULL)
}
\arguments{
\item{formula}{The formula call used by \code{survival::coxph(formula = Surv(time, dead) ~ x + y)}.
Do not include transformed variables in the formula (e.g. \code{Surv(time, dead) ~ log(x) + y})--the specified
\code{lambda} will no longer be interpretable.}

\item{data}{A data.frame or tibble with named columns with the columns specifed in formula.}

\item{pred.time}{The time at which the function will estimate the survival probabilites.}

\item{lambda}{The radius of the kernel for tri-cubic and Epanechnikov kernels.
The standard deviation for the Gaussian kernel.}

\item{grid}{A data.frame or tibble with named columns that are the covaraites specified in formula.
Survival probabilities will only be estimated at these points if specified.  Otherwise, survival probabilities
will be estimated at each unique combination of the observed covaraites.}
}
\value{
A data.frame or tibble with the covariates and the estiamted survival probability.
}
\description{
The purpose of this function is to proivde kernel-weighted n-year survival
estimates for n-dimensial data. For example, using two predictors, x and y,
predict time to death. The program estimates a hyperplane that estimates
risk of death at n-years for every x, y combination.
}
