% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.timeto.R
\name{sm.timeto}
\alias{sm.timeto}
\title{n-dimensional kernel-smoothed survival probabilities}
\usage{
sm.timeto(formula, data, lambda = 1, newdata = NULL, type = "survival",
  kernel = c("epanechnikov", "tricube", "gaussian", "knn"),
  dist.method = "euclidean", model = "coxph", knn = NULL,
  scale = as.logical(length(covars > 1)), details = FALSE)
}
\arguments{
\item{formula}{The formula call used by \code{survival::coxph(formula = Surv(time, dead) ~ x + y)}.
Do not include transformed variables in the formula (e.g. \code{Surv(time, dead) ~ log(x) + y})--the specified
\code{lambda} will no longer be interpretable.}

\item{data}{A data.frame or tibble with named columns with the columns specified in formula.}

\item{lambda}{The radius of the kernel for tri-cubic and Epanechnikov kernels.
The standard deviation for the Gaussian kernel.}

\item{newdata}{A data.frame or tibble with named columns that are the covariates specified in formula.
Survival probabilities will only be estimated at these points if specified.  Otherwise, survival probabilities
will be estimated at each unique combination of the observed covariates.}

\item{type}{Specifies the type of statistic that will be calculated.  Default is \code{survival}.}

\item{kernel}{Specifies the kernel to be used: \code{epanechnikov}, \code{tricube}, \code{gaussian}, and \code{knn} are accepted.
Default is \code{epanechnikov}.  See vignettes for differences in methods.}

\item{dist.method}{Specifies the distance measure to be used in the kernel.  Default is \code{euclidean}.
Any distance measure accepted by \code{stats::dist} is acceptable.}

\item{model}{Specifies the type of model that will be used for the estimation  Default is \code{coxph}.}

\item{knn}{Only used with \code{kernel == "knn"}.  Positive integer that specifies how many observations (k-nearest neighbors)
to include for each point estimate.}

\item{scale}{If TRUE, the covariates are scaled to mean 0 and standard deviation 1 prior to kernel smoothing.
For one covariate the default is scale = FALSE, and for 2 of more covariates the default is scale = TRUE.
If including 2 or more covariates, scale should be set to \code{TRUE}.}
}
\value{
A vector with the estimated survival probability.
}
\description{
The purpose of this function is to provide kernel-smoothed n-year survival
estimates for n-dimensional data. For example, using two predictors, x and y,
predict time to death. The program estimates a hyperplane that estimates
risk of death at n-years for every x, y combination.
}
