% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_predict.R, R/utils-sm_predict_raw.R
\name{sm_predict}
\alias{sm_predict}
\alias{sm_predict_raw_safely}
\title{Kernel-weighted regression estimation}
\usage{
sm_predict(data, method, formula, type, newdata = data,
  method.args = NULL, lambda = 1, verbose = FALSE)

sm_predict_raw_safely(...)
}
\arguments{
\item{data}{data frame}

\item{method}{function to use}

\item{formula}{formula}

\item{type}{type of statistic to smooth (e.g. survival, median survival, etc.)}

\item{newdata}{new data frame.  Default is `data`. Only requires covarites from
the RHS of `~` and the time compenent from the outcome for some survival estimators.}

\item{method.args}{List of additional arguments passed on to the
modelling function defined by `method`}

\item{lambda}{The radius of the kernel for tri-cubic, Epanechnikov, and flat kernels.
The standard deviation for the Gaussian kernel}

\item{verbose}{Return full set of results. Default is `FALSE`}
}
\description{
Provides smoothed estimates from a variety of models, but was built to work
primarily with time to event endpoints.
}
\keyword{internal}
