% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_predict.R
\name{sm_predict}
\alias{sm_predict}
\title{Kernel-weighted predictions from regression models}
\usage{
sm_predict(data, method, formula, type, newdata = data,
  method.args = NULL, lambda = 1, verbose = FALSE)
}
\arguments{
\item{data}{data frame}

\item{method}{function to use}

\item{formula}{formula}

\item{type}{type of statistic to smooth (e.g. survival, median survival, etc.)}

\item{newdata}{new data frame.  Default is `data`. Only requires covariates from
the RHS of `~` and the time component from the outcome for some survival estimators.}

\item{method.args}{List of additional arguments passed on to the
modelling function defined by `method`}

\item{lambda}{The radius of the kernel for tri-cubic, Epanechnikov, and flat kernels.
The standard deviation for the Gaussian kernel}

\item{verbose}{Return full set of results. Default is `FALSE`}
}
\description{
Calculates kernel-weighted predictions from regression models
(i.e. outcomes that can be calculated from the predict function).
}
\examples{
sm_predict(
  mtcars,
  method = "glm",
  formula = am ~ mpg,
  method.args = list(family = binomial(link = "logit")),
  type = "response"
)
}
