% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_regression.R
\name{sm_regression}
\alias{sm_regression}
\title{Kernel-weighted regression modelling}
\usage{
sm_regression(data, method, formula, weighting_var, newdata = data,
  method.args = NULL, lambda = 1, kernel = "epanechnikov",
  dist.method = "euclidean")
}
\arguments{
\item{data}{data frame}

\item{method}{function to use}

\item{formula}{formula}

\item{weighting_var}{columns name(s) of variables used to calculate weights}

\item{newdata}{new data frame.  Default is `data`.}

\item{method.args}{List of additional arguments passed on to the
modelling function defined by `method`}

\item{lambda}{The radius of the kernel for tri-cubic, Epanechnikov, and flat kernels.
The standard deviation for the Gaussian kernel}

\item{kernel}{Specifies the kernel to be used: `epanechnikov`, `tricube`,
`gaussian`, and `flat` are accepted. Default is `epanechnikov`}

\item{dist.method}{Specifies the distance measure to be used in the kernel.
Default is `euclidean`. Distance measures accepted by
\code{stats::\link[stats]{dist}} is acceptable.}
}
\description{
General function that fits weighted regerssion models, where the weights
are calculated from a ancillary variable(s) or from variable(s) found in the
regression model. Using `data`, `method`, and `formula` the regression model is
estimated; the estimates are weighted by the variable(s) listed in `weighting_var`.
}
\examples{
sm_regression(
  data = mtcars,
  method = "lm",
  formula = mpg ~ am ,
  weighting_var = "hp",
  lambda = 2
)
}
